+++ MLEM.Ui/Elements/SquishingGroup.cs
@@ -31,20 +31,6 @@ namespace MLEM.Ui.Elements {
         }
 
         /// <inheritdoc />
-        public override void Draw(GameTime time, SpriteBatch batch, float alpha, SpriteBatchContext context) {
-            // be sure to update our area before drawing since our children may not have been squished yet
-            this.UpdateAreaIfDirty();
-            base.Draw(time, batch, alpha, context);
-        }
-
-        /// <inheritdoc />
-        public override Element GetElementUnderPos(Vector2 position) {
-            // be sure to update our area before since our children may not have been squished yet
-            this.UpdateAreaIfDirty();
-            return base.GetElementUnderPos(position);
-        }
-
-        /// <inheritdoc />
         protected override void OnChildAreaDirty(Element child, bool grandchild) {
             base.OnChildAreaDirty(child, grandchild);
             if (!grandchild)